# allocates a thread for each store in /stores for speedup