--- a/shopsafe-backend/src/main/java/com/google/sps/servlets/StoresServlet.java
+++ b/shopsafe-backend/src/main/java/com/google/sps/servlets/StoresServlet.java
@@ -40,8 +40,6 @@ import java.util.Scanner;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.ThreadFactory;