-Creates a new thread for each nearby store in /stores
-Adds singular test for benchmarking speed

Results for philly which has 20 stores come up in places api call
Threading | Single
by Store    | Thread
------------- | ---------
7.113 sec  | 13.935 sec
6.704 sec  | 12.923 sec
6.44 sec    | 13.227 sec