@Raulp8 2020-07-20
After merging gabe's changes new time with same input:

Threading by Store and now using CSV
--
3.995 sec
3.303 sec
3.168 sec